requirements.txt
mplcursors==0.3
numpy==1.18.4
yfinance==0.1.54
tqdm==4.48.0
joblib==0.16.0
matplotlib==3.2.1
Flask==1.1.2
Flask_FlatPages==0.7.2
pandas==1.0.3
Frozen_Flask==0.15
python_dateutil==2.8.1
praw==7.1.0
python-dotenv==0.14.0
wheel==0.35.1
yahoo-finance-1.4.0
requests=2.22.0
